Add setContexts() method to prevent DMs to the bot
This patch adds the method setContexts() to the commands channel.js,
information.js and rules.js, to prevent users to use the commands
through DMs. The commands are restricted so only an administrator
can use it in the server. Despite these restrictions, the commands
could still be used by DMs from normal users, so this fixes the issue.