Add commands, fix some comments and minor fixes
The following commands were added:

- 'backup.js': creates a backup of the server, including channels,
permissions, roles, emojis and messages. However, it is subject to
change because it doesn't load the backup as expected.

- 'stealemoji.js': uploads a single emoji to a server.

- 'transferemojis.js: creates a backup of all the emojis from a server
in a JSON file, and it can be loaded to another server. If there's
more than 50 emojis to upload, the command will continue to run until
it uploads all the emojis. This happens because of the Discord API
rate limit.

A comment was fixed in the commands 'channels.js', 'information.js'
and 'rulesjs'.

The '.gitignore' file was modified so it disregard the 'backups' and
'emojis' directories.

Finally, the packages 'discord-backup' and 'node-fetch' were installed
for the 'backup.js' command.