Update discord.js, install sharp and add command
'discord.js' was updated from 14.16.3 to 14.9.3 so it can support
Components V2. 'sharp' was installed to resize images
programmatically.

The command 'braindump' was created. It serves as an index to redirect
via URL to various personal resources. It uses Components V2.

Finally, a new image was added for this command.