Migrate commands to Components V2
The following commands have been migrated to Components V2:

- information.js
- rules.js
- backup.js
- stealemoji.js
- transferemojis.js

The 'channels.js' command has been removed as it is now in the
'information.js' command.

Finally, the images have been renamed and moved to the container
directory for better organization, and the 'braindump.js' and 'rules.js'
commands use new images.